requirements.txt
minio==7.2.7
python-dotenv==1.0.1
urllib3==2.2.2